move async from mainDedupe() to query useEffect
